Manually backport items in renamed type.
AutomationParametersParameterParameters type in v1beta1 corresponds to
ParameterParameters type in v1beta2. It's not possible to "backport
the name" of the v1beta1 type to ParameterParameters, because that
name is already defined in Document resource (in zz_document_types.go
file) in v1beta1 package. Changes in such renamed types could be the
most difficult to detect and handle automatically.

Signed-off-by: Cem Mergenci <cmergenci@gmail.com>
Signed-off-by: Nick Atzert <nick.atzert@upbound.io>