update group name of namespaced core/config types to be aws.m.upbound.io
This isn't an ideal name because it doesn't exactly match the same pattern
of other groups, e.g. s3.m.aws.upbound.io, but it is a quick way to make
publishing work for the family/config provider. That provider is special
cased to take all aws.* CRDs. Updating the group for namespaced core/config
CRDs to aws.m.upbound.io means they are also included in the family/config
provider along with the cluster scoped aws.upbound.io CRDs. We should likely
revisit this with more time to maybe update the up xpkg batch tooling.

Signed-off-by: Jared Watts <jbw976@gmail.com>